wip - initial redux structure and migration of home
