+++ admin/client/actions/HomeActions.js
@@ -1,0 +1,8 @@
+import { SAVE_DATA } from '../constants/ActionTypes';
+
+export function saveData(text) {
+	return {
+		type: SAVE_DATA,
+		text
+	};
+}
