+++ admin/client/constants/ActionTypes.js
@@ -1,0 +1,4 @@
+export const DATA_FETCHED = 'DATA_FETCHED';
+export const DATA_SUCCEEDED = 'DATA_SUCCEEDED';
+export const DATA_FAILED = 'DATA_FAILED';
+export const SAVE_DATA = 'SAVE_DATA';
