+++ admin/client/packages.js
@@ -25,6 +25,10 @@ module.exports = [
 	'react-dom',
 	'react-select',
 	'react',
+	'redux',
+	'react-redux',
+	'redux-thunk',
+	'redux-devtools',
 	'store-prototype',
 	'superagent',
 	'vkey',
