+++ admin/client/reducers/index.js
@@ -1,0 +1,10 @@
+import { combineReducers } from 'redux';
+import Keystone from './Keystone';
+import Home from './Home';
+
+const rootReducer = combineReducers({
+	Keystone,
+	Home
+});
+
+export default rootReducer;
