+++ admin/client/stores/configureStore.js
@@ -1,0 +1,23 @@
+import { createStore, applyMiddleware, combineReducers, compose } from 'redux';
+import thunkMiddleware from 'redux-thunk';
+import { devTools, persistState } from 'redux-devtools';
+import * as reducers from '../reducers/index';
+
+let createStoreWithMiddleware;
+
+// Configure the dev tools when in DEV mode
+if (Keystone.devMode) {
+	createStoreWithMiddleware = compose(
+		applyMiddleware(thunkMiddleware),
+		devTools(),
+		persistState(window.location.href.match(/[?&]debug_session=([^&]+)\b/))
+	)(createStore);
+} else {
+	createStoreWithMiddleware = applyMiddleware(thunkMiddleware)(createStore);
+}
+
+const rootReducer = combineReducers(reducers);
+
+export default function configureStore(initialState) {
+	return createStoreWithMiddleware(rootReducer, initialState);
+}
